+++ utils/error/index.js
@@ -1,4 +1,4 @@
-/* Copyright (c) 2022, VRAI Labs and/or its affiliates. All rights reserved.
+/* Copyright (c) 2020, VRAI Labs and/or its affiliates. All rights reserved.
  *
  * This software is licensed under the Apache License, Version 2.0 (the
  * "License") as published by the Apache Software Foundation.
@@ -12,13 +12,9 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-
-/**
- * Refer to this issue to know why this is required: https://github.com/supertokens/supertokens-website/issues/134
- */
-export type CookieHandlerInterface = {
-    setCookie: (cookieString: string) => Promise<void>;
-    getCookie: () => Promise<string>;
-};
-
-export type CookieHandlerInput = (original: CookieHandlerInterface) => CookieHandlerInterface;
+"use strict";
+function __export(m) {
+  for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
+}
+exports.__esModule = true;
+__export(require("../../lib/build/error"));