+++ CHANGELOG.md
@@ -12,6 +12,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `API_INVALID_CLAIM` event and the related `invalidClaimStatusCode` config prop.
 - Adds base classes, types for session claims, and a `validateClaims` functions.
 
+## [13.0.1] - 2022-06-27
+- Fixes node dependency for integration tests
+
+## [13.0.0] - 2022-06-27
+
+### Breaking changes
+
+- Removes `setCookieSync` and `getCookieSync` from the interface for `cookieHandler` when calling `SuperTokens.init`
+
+## [12.2.0] - 2022-06-25
+- Adds new FDI support (1.14)
+
+## [12.1.0] - 2022-06-15
+
+### Changed
+
+- Calling `SuperTokens.signOut` now throws `STGeneralError` if the API responds with a general error status
+
 ## [12.0.0] - 2022-06-02
 
 - Adds tests for custom cookie and window handlers